Updates for many codes

+++ Factorial.cpp
@@ -2,10 +2,10 @@
 #include<math.h>
 using namespace std;
 int main(){
-    int n,nxt;
+    long int n,nxt;
     cout<<"Enter The Number for factorial\n";
     cin>>n;
-    int z=1;
+    long int z=1;
     for(int i=1;i<=n;++i){
         z*=i;
     }
